added inteerface for controller
Former-commit-id: 28535d888c3fe339bf72062500b4e0d4ad86ed3e

+++ src/gameengine/grid/interfaces/ActorGrid/ReadAndMoveGrid.java
@@ -2,6 +2,6 @@ package gameengine.grid.interfaces.ActorGrid;
 
 public interface ReadAndMoveGrid extends ReadableGrid{
 
-	void move(double ID, double newX, double newY);
+	void move(int ID, double newX, double newY);
 	
 }
